requirements.txt
requests
vk_api
anytree

PySocks